--- app/src/main/java/ru/jjba/jr2/data/adapter/WordAdapter.kt
@@ -1,9 +1,0 @@
-package ru.jjba.jr2.data.adapter
-
-import com.squareup.moshi.FromJson
-
-/*
-class WordAdapter {
-    @FromJson
-    fun fromJson
-}*/
