+++ app/src/main/java/ru/jjba/jr2/presentation/ui/gramma/GrammaFragment.kt
@@ -1,0 +1,22 @@
+package ru.jjba.jr2.presentation.ui.gramma
+
+import ru.jjba.jr2.presentation.ui.BaseFragment
+import ru.jjba.jr2.presentation.viewmodel.gramma.GrammaViewModel
+
+class GrammaFragment : BaseFragment<GrammaViewModel>() {
+    override val layoutRes: Int
+        get() = TODO("not implemented")
+    override val titleDefault: String
+        get() = TODO("not implemented")
+    override var viewModel: GrammaViewModel
+        get() = TODO("not implemented")
+        set(value) {}
+
+    override fun initContent() {
+
+    }
+
+    override fun observeData() {
+
+    }
+}
