--- app/src/main/java/ru/jjba/jr2/presentation/ui/main/activity/MainActivity.kt
@@ -1,23 +1,0 @@
-package ru.jjba.jr2.presentation.ui.main.activity
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.NavigationUI
-import kotlinx.android.synthetic.main.activity_main.*
-import ru.jjba.jr2.R
-
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setupBottomNavigation()
-    }
-
-    private fun setupBottomNavigation() {
-        val navHostFragment = supportFragmentManager
-                .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        NavigationUI.setupWithNavController(bottomNavigationView, navHostFragment.navController)
-    }
-}
