+++ app/src/main/java/ru/jjba/jr2/presentation/ui/vocab/VocabFragment.kt
@@ -1,0 +1,22 @@
+package ru.jjba.jr2.presentation.ui.vocab
+
+import ru.jjba.jr2.presentation.ui.BaseFragment
+import ru.jjba.jr2.presentation.viewmodel.vocab.VocabViewModel
+
+class VocabFragment : BaseFragment<VocabViewModel>() {
+    override val layoutRes: Int
+        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
+    override val titleDefault: String
+        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
+    override var viewModel: VocabViewModel
+        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
+        set(value) {}
+
+    override fun initContent() {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun observeData() {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
