+++ app/src/main/java/ru/jjba/jr2/presentation/viewmodel/gramma/GrammaViewModel.kt
@@ -1,0 +1,7 @@
+package ru.jjba.jr2.presentation.viewmodel.gramma
+
+import androidx.lifecycle.ViewModel
+
+class GrammaViewModel : ViewModel() {
+
+}
