--- app/src/main/java/ru/jjba/jr2/presentation/viewmodel/interpritation/InterpretationListViewModel.kt
@@ -1,7 +1,0 @@
-package ru.jjba.jr2.presentation.viewmodel.interpritation
-
-import androidx.lifecycle.ViewModel
-
-class InterpretationListViewModel : ViewModel() {
-
-}
