--- app/src/main/java/ru/jjba/jr2/presentation/viewmodel/kana/details/KanaDetailsViewModel.kt
@@ -1,6 +1,0 @@
-package ru.jjba.jr2.presentation.viewmodel.kana.details
-
-import androidx.lifecycle.ViewModel
-
-class KanaDetailsViewModel : ViewModel() {
-}
