--- app/src/main/java/ru/jjba/jr2/presentation/viewmodel/kana/list/KanaListViewModel.kt
@@ -1,6 +1,0 @@
-package ru.jjba.jr2.presentation.viewmodel.kana.list
-
-import androidx.lifecycle.ViewModel
-
-class KanaListViewModel : ViewModel() {
-}
