+++ app/src/main/java/ru/jjba/jr2/presentation/viewmodel/profile/ProfileViewModel.kt
@@ -1,0 +1,6 @@
+package ru.jjba.jr2.presentation.viewmodel.profile
+
+import androidx.lifecycle.ViewModel
+
+class ProfileViewModel : ViewModel() {
+}
