+++ app/src/main/java/ru/jjba/jr2/presentation/viewmodel/vocab/VocabViewModel.kt
@@ -1,0 +1,6 @@
+package ru.jjba.jr2.presentation.viewmodel.vocab
+
+import androidx.lifecycle.ViewModel
+
+class VocabViewModel : ViewModel() {
+}
