+++ app/src/main/java/ru/jjba/jr2/presentation/viewmodel/vocab/kana/KanaListViewModel.kt
@@ -1,0 +1,6 @@
+package ru.jjba.jr2.presentation.viewmodel.vocab.kana
+
+import androidx.lifecycle.ViewModel
+
+class KanaListViewModel : ViewModel() {
+}
