+++ app/src/main/java/ru/jjba/jr2/presentation/viewmodel/vocab/word/WordDetailViewModel.kt
@@ -1,0 +1,6 @@
+package ru.jjba.jr2.presentation.viewmodel.vocab.word
+
+import androidx.lifecycle.ViewModel
+
+class WordDetailViewModel : ViewModel() {
+}
