--- app/src/main/java/ru/jjba/jr2/presentation/viewmodel/word/details/WordDetailsViewModel.kt
@@ -1,6 +1,0 @@
-package ru.jjba.jr2.presentation.viewmodel.word.details
-
-import androidx.lifecycle.ViewModel
-
-class WordDetailsViewModel : ViewModel() {
-}
