+++ app/src/main/res/layout/fragment_kana_detail.xml
