+++ app/src/main/res/layout/fragment_word_detail.xml
