+++ app/src/main/res/menu/bottom_navigation.xml
@@ -6,11 +6,15 @@
         android:icon="@drawable/ic_home"
         android:title="@string/menu_main_title" />
     <item
-        android:id="@+id/navigation_word_list"
+        android:id="@+id/navigation_vocab"
         android:icon="@drawable/ic_word_list"
-        android:title="@string/menu_word_list_title" />
+        android:title="@string/menu_vocab_title" />
     <item
-        android:id="@+id/navigation_kana"
+        android:id="@+id/navigation_gramma"
         android:icon="@drawable/ic_word_list"
-        android:title="@string/menu_kana_title" />
+        android:title="@string/menu_gramma_title" />
+    <item
+        android:id="@+id/navigation_profile"
+        android:icon="@drawable/ic_word_list"
+        android:title="@string/menu_profile_title" />
 </menu>
