+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/navigation_main"
-        android:name="ru.jjba.jr2.presentation.ui.main.fragment.MainFragment"
+        android:name="ru.jjba.jr2.presentation.ui.main.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" />
     <fragment
         android:id="@+id/navigation_word_list"
-        android:name="ru.jjba.jr2.presentation.ui.word.list.WordListFragment"
+        android:name="ru.jjba.jr2.presentation.ui.vocab.word.list.WordListFragment"
         android:label="WordListFragment"
         tools:layout="@layout/fragment_word_list" />
     <fragment
         android:id="@+id/navigation_kana"
-        android:name="ru.jjba.jr2.presentation.ui.kana.list.KanaListFragment"
+        android:name="ru.jjba.jr2.presentation.ui.vocab.kana.KanaListFragment"
         android:label="KanaFragment"
         tools:layout="@layout/fragment_kana_list" />
 </navigation>
